Fix #29881 and fix #26736, as well as other UI adjustments
We've moved the minimum size of 600 out of the UI and instead set it via splitter->setSizes(),
which allows us to have an eventual horizontal scrollbar for the right-hand side panel;
as a side-effect, the vertical buttons no longer overlap with tree views when the width is too small.

We also make sure that when hiding the Form Layout panel (by switching the form designer mode),
we have nice alignment between the search box horizontal layout, the Available Widgets panel,
and the vertical buttons (which by the way, now won't claim too much horizontal space when Form
Layout is not present).

Lots of UI tweaks :)